+++ database/migrations/2017_01_15_000000_add_permission_group_id_to_permissions_table.php
@@ -25,8 +25,10 @@ class AddPermissionGroupIdToPermissionsTable extends Migration
      */
     public function down()
     {
-        Schema::table('permissions', function (Blueprint $table) {
-            $table->dropColumn('permission_group_id');
-        });
+        if (!App::runningUnitTests()) {
+            Schema::table('permissions', function (Blueprint $table) {
+                $table->dropColumn('permission_group_id');
+            });
+        }
     }
 }
