+++ database/migrations/2017_08_05_000000_add_group_to_settings_table.php
@@ -25,8 +25,10 @@ class AddGroupToSettingsTable extends Migration
      */
     public function down()
     {
-        Schema::table('settings', function (Blueprint $table) {
-            $table->dropColumn('group');
-        });
+        if (!App::runningUnitTests()) {
+            Schema::table('settings', function (Blueprint $table) {
+                $table->dropColumn('group');
+            });
+        }
     }
 }
